Always send log error emails and implement "lockfile"
Changes runner bash script to move "send
email" code to the exit trap, so even if the
python code errors out, the log file will still be
checked for "bad" strings and an email will be
sent (resolves one of the problems of #100).

Also implements "lock file checking". The script
will create a lock file (named ".builder.lock") while
it is running and remove it when finished. If this
file exists at the start of the script, the record
builder will not be run. This should avoid the
situation observed in #101